Optimize SD reads
- 32 bit memory accesses are a shitload faster than 8 bit
- Hardware stalling reads instead of software busy loop

Benchmark shows ~368KiB/s for sector read + copy to a second buffer

+++ firmware/src/sd.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 
 void sd_init(void);
-void sd_read(uint32_t lba, uint8_t *buf);
+void sd_read(uint32_t lba, uint32_t *buf);
 
 #endif
